backend/requirements.txt
celery==5.6.0
django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==25.0.0
psycopg[binary]==3.2.10
redis==6.4.0
